requirements.txt
fastapi==0.95.2
uvicorn==0.20.0
sqlalchemy==2.0.3
psycopg2-binary==2.9.5
redis==4.3.4
celery==5.2.7
pydantic==1.10.2
python-dotenv==1.0.0
alembic==1.9.4
pytest==7.2.1
httpx==0.24.0
pillow==9.5.0  # For image processing (if needed)
